Let ovn db cluster leaders start first
It looks like if one of the db nodes not designated as the initial
leader starts up first, the cluster starts in a bad state, leaving 2
of the 3 nodes not functional.

This specific change is a quick hack, just making the other 2 wait 5
more minutes before starting up.  If this ends up working, it will
need to be redone to use a proper check by connecting to the db.